backend/requirements.txt
flask
requests
celery
sqlalchemy
flask_cors
python-dotenv
pymysql